+++ encode.go
@@ -5,9 +5,28 @@ import (
 	"encoding/json"
 	"errors"
 	"io"
+	"io/fs"
 	"os"
 	"path/filepath"
 )
+
+// A CreateFS provides access to a hierarchical file system.
+// Must follow the same naming convention as io/fs.FS.
+type CreateFS interface {
+	fs.FS
+	Create(name string) (io.WriteCloser, error)
+}
+
+// dirFS implements a file system (an fs.FS) for the tree of files rooted at the directory dir.
+type dirFS struct {
+	fs.FS
+	dir string
+}
+
+// Create creates or truncates the named file.
+func (d dirFS) Create(name string) (io.WriteCloser, error) {
+	return os.Create(d.dir + "/" + name)
+}
 
 // Save will save a document as a glTF with the specified by name.
 func Save(doc *Document, name string) error {
